requirements.txt
python-telegram-bot==20.7
asyncpg==0.29.0
python-dotenv==1.0.0
telethon==1.34.0
cryptg==0.4.0
pillow==10.1.0
matplotlib==3.8.2
psutil==5.9.6
aiofiles==23.2.1
aiohttp==3.9.1
ujson==5.8.0
redis==5.0.1
schedule==1.2.0
